Moved generator transform methods to extensions
